move dense_xw_product_function to eval/instruction
