+++ eval/src/tests/tensor/dense_xw_product_function/dense_xw_product_function_test.cpp
@@ -6,7 +6,7 @@
 #include <vespa/eval/eval/simple_tensor.h>
 #include <vespa/eval/eval/simple_tensor_engine.h>
 #include <vespa/eval/tensor/default_tensor_engine.h>
-#include <vespa/eval/tensor/dense/dense_xw_product_function.h>
+#include <vespa/eval/instruction/dense_xw_product_function.h>
 #include <vespa/eval/tensor/dense/dense_tensor.h>
 #include <vespa/eval/tensor/dense/dense_tensor_view.h>
 #include <vespa/eval/eval/test/tensor_model.hpp>
