+++ eval/src/vespa/eval/tensor/default_tensor_engine.cpp
@@ -9,7 +9,7 @@
 #include "dense/dense_tensor.h"
 #include "dense/typed_dense_tensor_builder.h"
 #include "dense/dense_dot_product_function.h"
-#include "dense/dense_xw_product_function.h"
+#include <vespa/eval/instruction/dense_xw_product_function.h>
 #include "dense/dense_matmul_function.h"
 #include "dense/dense_multi_matmul_function.h"
 #include "dense/dense_fast_rename_optimizer.h"
